refactor: lib
