+++ src/lib/ghostscript/background-worker.js
