+++ src/lib/ghostscript/gs-worker.js
