+++ src/lib/ghostscript/worker-init.js
